x86: Increase per-CPU stack/data area to 12KiB
Increase the size of the per-CPU stack/data area from one page to three,
i.e. from 4KiB to 12KiB.  KVM-Unit-Tests currently places the per-CPU data
at the bottom of the stack page, i.e. the stack "page" is actually a page
minus the size of the per-CPU area.  And of course there's no guard page
or buffer in between the two, and so overflowing the stack clobbers per-CPU
data and sends tests into the weeds in weird ways.

Punt on less awful infrastructure, and settle for fixing the most egregious
problem of tests having less than 4KiB of stack to work with.

Link: https://lore.kernel.org/r/20250215012032.1206409-4-seanjc@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>