riscv: sbi: Drop fwft upper bits test
The test should be checking for SBI_SUCCESS and that the behavior of
feature = BIT(32) is the same as feature = 0 (MISALIGNED_EXC_DELEG).
However, enabling MISALIGNED_EXC_DELEG doesn't always lead to traps
in S-mode (the platform may support misaligned accesses and not
trap at all). Drop this upper bits of feature ID test for now.
We'll add it back with another feature that has a deterministic
behavior (such as with the ADUE feature).

Reviewed-by: Clément Léger <cleger@rivosinc.com>
Signed-off-by: Andrew Jones <andrew.jones@linux.dev>