x86: Delete split IRQ chip variants of apic and ioapic tests
Remove the dedicated {io,}apic-split testcases and instead rely on users
to run KVM-Unit-Tests with ACCEL="kvm,kernel_irqchip=split" to provide the
desired coverage.

While providing more coverage by "default" is nice, the flip side of doing
so is that makes it annoying for an end user to do more, and gives the
false impression that the configurations in unittests.cfg are the only
configurations worth testing.

E.g. with kernel_irqchip=split, svm_npt fails on x2AVIC hardware due to
test bugs, hyperv_connections fails due to what is effectively a QEMU bug
that also got hoisted into KVM], and vmx_apic_passthrough_tpr_threshold_test
fails (with some KVM versions) due to a KVM bug that happened to be masked
by another KVM bug with the in-kernel PIT emulation.

Link: https://lore.kernel.org/all/Z8ZBzEJ7--VWKdWd@google.com
Link: https://lore.kernel.org/all/202502271500.28201544-lkp@intel.com
Link: https://lore.kernel.org/all/20250304211223.124321-1-seanjc@google.com
Cc: Naveen N Rao <naveen@kernel.org>
Link: https://lore.kernel.org/r/20250604000812.199087-1-seanjc@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>