x86: Move call to load_idt() out of setup_tr_and_percpu macro
Move the call to load_idt() out of setup_tr_and_percpu and into the two
"callers" of the macro.  This will allow moving the BSP's invocation of
load_idt() into setup_idt(), without creating a confusing scenario where
load_idt() is called multiple times, e.g. so that it's clear the BSP is
responsible for configuring the IDT.

No functional change intended.

Link: https://lore.kernel.org/r/20250604183623.283300-4-seanjc@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>