x86: Load IDT on BSP as part of setup_idt()
Load the IDT on the BSP as part of setup_idt(), to guarantee that the IDT
is loaded when setup_idt() runs (currently, the EFI boot path loads the
IDT _after_ setup_idt().  This will allow probing for forced emulation,
which requires being able to handle a #UD, during setup_idt().

Link: https://lore.kernel.org/r/20250604183623.283300-5-seanjc@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>