nVMX: Force emulation of LGDT/LIDT in iff FEP is available
Force emulation of LGDT/LIDT in the canonical host values test if and only
if forced emulation is available, otherwise the resulting #UD will result
in failures.

Note, use of the cached is_fep_available is critical for the LIDT test as
the "slow" check for FEP support requires being able to eat a #UD, i.e.
would fail when the current IDT is garbage (loaded with one of the
arbitrary canonical test values).

Reported-by: Chenyi Qiang <chenyi.qiang@intel.com>
Link: https://lore.kernel.org/r/20250604183623.283300-7-seanjc@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>