x86: nSVM: Set MSRPM bit on-demand when testing interception
Set the to-be-tested {RD,WR}MSR's associated MSRPM bit on-demand when
verifying KVM's emulation of MSR interception.  Setting all bits in the
bitmap creates a massive blind spot in the test coverage, e.g. KVM could
completely ignore the bitmap, and the test would still pass.

Run the test with all bits initially clear, and all bits initially set,
mostly in prepartion for adding coverage of MSR passthrough in a future
commit.

Link: https://lore.kernel.org/r/20250605192226.532654-7-seanjc@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>