riscv: sbi: sse: Fix wrong sse stack initialization
The sse stack temporary storage wasn't correctly initialize to 0. Since
this is used to know if stack needs to be deallocated at the end (in
the error handling path), sse_stack_free() would crash if an error
happened while registering the sse event for instance. Fix this by
correctly initializing all the stacks entries.

Fixes: 7536c9add92b ("riscv: sbi: Add SSE extension tests")
Signed-off-by: Clément Léger <cleger@rivosinc.com>
Signed-off-by: Andrew Jones <andrew.jones@linux.dev>