riscv: lib: Pass test status on sbi shutdown
The SBI SRST system reset function used to exit the test takes two
parameters. The first is used to request a shutdown and the second
describes the reason. A value of zero means "no reason" and a value
of one means "system failure". Other values are currently reserved.
Use the value of zero to indicate that the tests completed and all
passed and of one otherwise.

It may be possible for the SBI implemention to use this information
to alert testers of failing or aborting tests.

Signed-off-by: Jesse Taube <jesse@rivosinc.com>
[Enhanced commit message and changed sbi_shutdown param to a bool.]
Signed-off-by: Andrew Jones <andrew.jones@linux.dev>