scripts: Add 'test_args' test definition parameter
kvm-unit-tests, on arm and arm64, is getting ready to support running all
the test automatically under kvmtool. Even though kvmtool has a different
syntax for configuring and running a virtual machine, kvmtool and qemu have
in common the test arguments that are passed to the main() function.

Add a new test definition parameter, 'test_args', that contains only the
VMM-independent arguments that are passed to the main() function, with the
intention for the parameter to be used by both qemu and kvmtool, when
support for kvmtool is added.

Suggested-by: Andrew Jones <andrew.jones@linux.dev>
Reviewed-by: Andrew Jones <andrew.jones@linux.dev>
Reviewed-by: Shaoqin Huang <shahuang@redhat.com>
Signed-off-by: Alexandru Elisei <alexandru.elisei@arm.com>
Signed-off-by: Andrew Jones <andrew.jones@linux.dev>