x86/run: Specify "-display none" instead of "-vnc none"
"-display none" is a more generic option that is always available,
unlike "-vnc none", so use it instead of probing for the existence
of -vnc.

This mostly reverts commit 0f982a8c1e2242483a9bf53b15f825d1ff0bccc6,
"x86/run: Specify "-vnc none" for QEMU if and only if QEMU supports VNC",
though without reintroducing the bug and with some conflicts that
happened in the meanwhile.

Reviewed-by: Sean Christopherson <seanjc@google.com>
Tested-by: Sean Christopherson <seanjc@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>