shellcheck: suppress SC2327,2328 false positives
Shellcheck warnings SC2327,SC2328 complain that a command substitution
will be empty if the output is redirected, which is a valid warning but
shellcheck is not smart enough to see when output is redirected into a
command that outputs what the command substitution wanted.

Add comments and shellcheck directives to these cases.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Andrew Jones <andrew.jones@linux.dev>