x86/vmexit: Add WBINVD and INVD VM-Exit latency testcases
Add WBINVD and INVD testcase to the VM-Exit performance/latency test so
that it's easy to measure latency of VM-Exits that are handled in KVM's
fastpath on both Intel and AMD (INVD), and so that a direct comparison can
be made to an exit with no meaningful emulation (WBINVD).

Don't create entries in x86/unittests.cfg, as running the INVD test on
bare metal (or a hypervisor that emulates INVD) would likely corrupt
memory (and similarly, WBINVD can have a massively negative impact on the
system).

Tested-by: Manali Shukla <manali.shukla@amd.com>
Link: https://patch.msgid.link/20251113235946.1710922-1-seanjc@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>